Move cookies to the lib
